Update to JVM SDK M25, closes #242

+++ it/controllers/ApplicationControllerIntegrationTest.java
@@ -17,7 +17,7 @@ public class ApplicationControllerIntegrationTest extends WithPlayJavaSphereClie
     @Test
     public void itFindsSomeCategories() throws Exception {
         final PagedQueryResult<Category> result = execute(CategoryQuery.of()).get(4000);
-        final int count = result.size();
+        final long count = result.size();
         assertThat(count).isGreaterThan(3);
         //this is a project specific assertion as example
         assertThat(toEnglishNames(result)).contains("Tank tops");
